\name{celsius2Kelvin}
\alias{celsius2Kelvin}
\title{celsius2Kelvin}
\description{Convert temperature from degrees of Celsius to Kelvin.}
\usage{celsius2Kelvin(temp)}
\arguments{
  \item{temp}{numeric vector: the temperature in degress of Celsius.}
}

\value{numeric vector of length of argument temp: Temperature in Kelvin.}

\author{Thomas Wutzler}





