\name{respTempLlyodAndTaylor94}
\alias{respTempLlyodAndTaylor94}
\title{respTempLlyodAndTaylor94}

\usage{
respTempLlyodAndTaylor94(R10, temp)
}
\arguments{
  \item{R10}{numeric vector: respiration rate 10 degrees of Celsius (usually a scalar parameter)}
  \item{temp}{
}
}
\details{Describes soil respiration at soil core scale. }

\references{Lloyd, J. & Taylor, J. (1994) On the temperature dependence of soil respiration. Functional ecology, JSTOR, 8, 315-323}
\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





