\name{showCppHelloWorld}
\alias{showCppHelloWorld}
\title{showCppHelloWorld}
\description{call internal function \code{.rcpp_hello_world}}
\usage{showCppHelloWorld(...)}
\arguments{
  \item{\dots}{further arguments to \code{.rcpp_hello_world}}
}
\details{This function demonstrates renaming of cpp funcitons.
Due to starting with the dot, it will not be available from outside
the package}
\value{result of \code{.rcpp_hello_world}}

\author{Thomas Wutzler}





